Refactor out QuadrantMap
