+++ src/api/geo.js
@@ -30,14 +30,18 @@ exports.addressHandler = async (req, res, next) => {
     const geoResults = geoResult.data.results;
 
     const locResult = geoResults[0];
+    const {
+      geometry: { location }
+    } = locResult;
+
     const neighborhood = locResult.address_components.find(component =>
       component.types.includes("neighborhood")
     );
+
     const neighborhoodName = neighborhood ? neighborhood.long_name : "";
-    const [lt, long] = [
-      locResult.geometry.location.lat,
-      locResult.geometry.location.lng
-    ];
+
+    const [lt, long] = [location.lat, location.lng];
+
     const intersection = await geonamesClient.findNearestIntersection({
       lat: lt,
       lng: long
@@ -53,6 +57,7 @@ exports.addressHandler = async (req, res, next) => {
     return res.end(
       JSON.stringify({
         neighborhoodName,
+        location,
         intersection: {
           street_1: intersection.intersection.street1,
           street_2: intersection.intersection.street2
